@@ -44,7 +44,7 @@ void RpcUtils::verifySuccess(const TSStatus &status) {
     }
 }
 
-void RpcUtils::verifySuccess(const vector <TSStatus> &statuses) {
+void RpcUtils::verifySuccess(const vector<TSStatus> &statuses) {
     for (TSStatus status: statuses) {
         if (status.code != TSStatusCode::SUCCESS_STATUS) {
             throw BatchExecutionException(statuses, status.message);
@@ -65,37 +65,37 @@ TSStatus RpcUtils::getStatus(int code, const string &message) {
     return status;
 }
 
-shared_ptr <TSExecuteStatementResp> RpcUtils::getTSExecuteStatementResp(TSStatusCode::TSStatusCode tsStatusCode) {
+shared_ptr<TSExecuteStatementResp> RpcUtils::getTSExecuteStatementResp(TSStatusCode::TSStatusCode tsStatusCode) {
     TSStatus status = getStatus(tsStatusCode);
     return getTSExecuteStatementResp(status);
 }
 
-shared_ptr <TSExecuteStatementResp>
+shared_ptr<TSExecuteStatementResp>
 RpcUtils::getTSExecuteStatementResp(TSStatusCode::TSStatusCode tsStatusCode, const string &message) {
     TSStatus status = getStatus(tsStatusCode, message);
     return getTSExecuteStatementResp(status);
 }
 
-shared_ptr <TSExecuteStatementResp> RpcUtils::getTSExecuteStatementResp(const TSStatus &status) {
-    shared_ptr <TSExecuteStatementResp> resp(new TSExecuteStatementResp());
+shared_ptr<TSExecuteStatementResp> RpcUtils::getTSExecuteStatementResp(const TSStatus &status) {
+    shared_ptr<TSExecuteStatementResp> resp(new TSExecuteStatementResp());
     TSStatus tsStatus(status);
     resp->status = status;
     return resp;
 }
 
-shared_ptr <TSFetchResultsResp> RpcUtils::getTSFetchResultsResp(TSStatusCode::TSStatusCode tsStatusCode) {
+shared_ptr<TSFetchResultsResp> RpcUtils::getTSFetchResultsResp(TSStatusCode::TSStatusCode tsStatusCode) {
     TSStatus status = getStatus(tsStatusCode);
     return getTSFetchResultsResp(status);
 }
 
-shared_ptr <TSFetchResultsResp>
+shared_ptr<TSFetchResultsResp>
 RpcUtils::getTSFetchResultsResp(TSStatusCode::TSStatusCode tsStatusCode, const string &appendMessage) {
     TSStatus status = getStatus(tsStatusCode, appendMessage);
     return getTSFetchResultsResp(status);
 }
 
-shared_ptr <TSFetchResultsResp> RpcUtils::getTSFetchResultsResp(const TSStatus &status) {
-    shared_ptr <TSFetchResultsResp> resp(new TSFetchResultsResp());
+shared_ptr<TSFetchResultsResp> RpcUtils::getTSFetchResultsResp(const TSStatus &status) {
+    shared_ptr<TSFetchResultsResp> resp(new TSFetchResultsResp());
     TSStatus tsStatus(status);
     resp->__set_status(tsStatus);
     return resp;
@@ -103,6 +103,10 @@ shared_ptr <TSFetchResultsResp> RpcUtils::getTSFetchResultsResp(const TSStatus &
 
 void Tablet::reset() {
     rowSize = 0;
+    for (int i = 0; i < schemas.size(); i++) {
+        BitMap *bitMap = bitMaps[i].get();
+        bitMap->reset();
+    }
 }
 
 void Tablet::createColumns() {
@@ -168,30 +172,51 @@ string SessionUtils::getValue(const Tablet &tablet) {
     MyStringBuffer valueBuffer;
     for (size_t i = 0; i < tablet.schemas.size(); i++) {
         TSDataType::TSDataType dataType = tablet.schemas[i].second;
+        BitMap *bitMap = tablet.bitMaps[i].get();
         switch (dataType) {
             case TSDataType::BOOLEAN:
                 for (int index = 0; index < tablet.rowSize; index++) {
-                    valueBuffer.putBool(tablet.values[i][index] == "true");
+                    if (!bitMap->isMarked(index)) {
+                        valueBuffer.putBool(tablet.values[i][index] == "true");
+                    } else {
+                        valueBuffer.putBool(false);
+                    }
                 }
                 break;
             case TSDataType::INT32:
                 for (int index = 0; index < tablet.rowSize; index++) {
-                    valueBuffer.putInt(stoi(tablet.values[i][index]));
+                    if (!bitMap->isMarked(index)) {
+                        valueBuffer.putInt(stoi(tablet.values[i][index]));
+                    } else {
+                        valueBuffer.putInt((numeric_limits<int>::min)());
+                    }
                 }
                 break;
             case TSDataType::INT64:
                 for (int index = 0; index < tablet.rowSize; index++) {
-                    valueBuffer.putLong(stol(tablet.values[i][index]));
+                    if (!bitMap->isMarked(index)) {
+                        valueBuffer.putLong(stol(tablet.values[i][index]));
+                    } else {
+                        valueBuffer.putLong((numeric_limits<int64_t>::min)());
+                    }
                 }
                 break;
             case TSDataType::FLOAT:
                 for (int index = 0; index < tablet.rowSize; index++) {
-                    valueBuffer.putFloat(stof(tablet.values[i][index]));
+                    if (!bitMap->isMarked(index)) {
+                        valueBuffer.putFloat(stof(tablet.values[i][index]));
+                    } else {
+                        valueBuffer.putFloat((numeric_limits<float>::min)());
+                    }
                 }
                 break;
             case TSDataType::DOUBLE:
                 for (int index = 0; index < tablet.rowSize; index++) {
-                    valueBuffer.putDouble(stod(tablet.values[i][index]));
+                    if (!bitMap->isMarked(index)) {
+                        valueBuffer.putDouble(stod(tablet.values[i][index]));
+                    } else {
+                        valueBuffer.putDouble((numeric_limits<double>::min)());
+                    }
                 }
                 break;
             case TSDataType::TEXT:
@@ -201,6 +226,17 @@ string SessionUtils::getValue(const Tablet &tablet) {
                 break;
             default:
                 throw UnSupportedDataTypeException(string("Data type ") + to_string(dataType) + " is not supported.");
+        }
+    }
+    for (size_t i = 0; i < tablet.schemas.size(); i++) {
+        BitMap *bitMap = tablet.bitMaps[i].get();
+        bool columnHasNull = !bitMap->isAllUnmarked();
+        valueBuffer.putChar(columnHasNull ? (char) 1 : (char) 0);
+        if (columnHasNull) {
+            vector<char> bytes = bitMap->getByteArray();
+            for (int j = 0; j < bytes.size(); j++) {
+                valueBuffer.putChar(bytes[j]);
+            }
         }
     }
     return valueBuffer.str;
@@ -214,21 +250,21 @@ void SessionDataSet::setBatchSize(int batchSize) {
     this->batchSize = batchSize;
 }
 
-vector <string> SessionDataSet::getColumnNames() { return this->columnNameList; }
+vector<string> SessionDataSet::getColumnNames() { return this->columnNameList; }
 
 bool SessionDataSet::hasNext() {
     if (hasCachedRecord) {
         return true;
     }
     if (!tsQueryDataSetTimeBuffer.hasRemaining()) {
-        shared_ptr <TSFetchResultsReq> req(new TSFetchResultsReq());
+        shared_ptr<TSFetchResultsReq> req(new TSFetchResultsReq());
         req->__set_sessionId(sessionId);
         req->__set_statement(sql);
         req->__set_fetchSize(batchSize);
         req->__set_queryId(queryId);
         req->__set_isAlign(true);
         try {
-            shared_ptr <TSFetchResultsResp> resp(new TSFetchResultsResp());
+            shared_ptr<TSFetchResultsResp> resp(new TSFetchResultsResp());
             client->fetchResults(*resp, *req);
             RpcUtils::verifySuccess(resp->status);
 
@@ -252,7 +288,7 @@ bool SessionDataSet::hasNext() {
 }
 
 void SessionDataSet::constructOneRow() {
-    vector <Field> outFields;
+    vector<Field> outFields;
     int loc = 0;
     for (int i = 0; i < columnSize; i++) {
         Field field;
@@ -313,7 +349,12 @@ void SessionDataSet::constructOneRow() {
         outFields.push_back(field);
     }
 
-    rowRecord = RowRecord(tsQueryDataSetTimeBuffer.getLong(), outFields);
+    if (!this->isIgnoreTimeStamp) {
+        rowRecord = RowRecord(tsQueryDataSetTimeBuffer.getLong(), outFields);
+    } else {
+        tsQueryDataSetTimeBuffer.getLong();
+        rowRecord = RowRecord(outFields);
+    }
     rowsIndex++;
 }
 
@@ -335,11 +376,11 @@ RowRecord *SessionDataSet::next() {
 }
 
 void SessionDataSet::closeOperationHandle() {
-    shared_ptr <TSCloseOperationReq> closeReq(new TSCloseOperationReq());
+    shared_ptr<TSCloseOperationReq> closeReq(new TSCloseOperationReq());
     closeReq->__set_sessionId(sessionId);
     closeReq->__set_statementId(statementId);
     closeReq->__set_queryId(queryId);
-    shared_ptr <TSStatus> closeResp(new TSStatus());
+    shared_ptr<TSStatus> closeResp(new TSStatus());
     try {
         client->closeOperation(*closeResp, *closeReq);
         RpcUtils::verifySuccess(*closeResp);
@@ -371,7 +412,7 @@ bool Session::checkSorted(const Tablet &tablet) {
     return true;
 }
 
-bool Session::checkSorted(const vector <int64_t> &times) {
+bool Session::checkSorted(const vector<int64_t> &times) {
     for (size_t i = 1; i < times.size(); i++) {
         if (times[i] < times[i - 1]) {
             return false;
@@ -400,8 +441,8 @@ void Session::sortTablet(Tablet &tablet) {
     delete[] index;
 }
 
-void Session::sortIndexByTimestamp(int *index, std::vector <int64_t> &timestamps, int length) {
-    if ( length <= 1 ) {
+void Session::sortIndexByTimestamp(int *index, std::vector<int64_t> &timestamps, int length) {
+    if (length <= 1) {
         return;
     }
 
@@ -431,7 +472,7 @@ void Session::appendValues(string &buffer, const char *value, int size) {
 }
 
 void
-Session::putValuesIntoBuffer(const vector <TSDataType::TSDataType> &types, const vector<char *> &values, string &buf) {
+Session::putValuesIntoBuffer(const vector<TSDataType::TSDataType> &types, const vector<char *> &values, string &buf) {
     for (size_t i = 0; i < values.size(); i++) {
         int8_t typeNum = getDataTypeNumber(types[i]);
         buf.append((char *) (&typeNum), sizeof(int8_t));
@@ -496,7 +537,7 @@ void Session::open(bool enableRPCCompression, int connectionTimeoutInMs) {
         return;
     }
 
-    shared_ptr <TSocket> socket(new TSocket(host, rpcPort));
+    shared_ptr<TSocket> socket(new TSocket(host, rpcPort));
     transport = std::make_shared<TFramedTransport>(socket);
     socket->setConnTimeout(connectionTimeoutInMs);
     if (!transport->isOpen()) {
@@ -508,10 +549,10 @@ void Session::open(bool enableRPCCompression, int connectionTimeoutInMs) {
         }
     }
     if (enableRPCCompression) {
-        shared_ptr <TCompactProtocol> protocol(new TCompactProtocol(transport));
+        shared_ptr<TCompactProtocol> protocol(new TCompactProtocol(transport));
         client = std::make_shared<TSIServiceClient>(protocol);
     } else {
-        shared_ptr <TBinaryProtocol> protocol(new TBinaryProtocol(transport));
+        shared_ptr<TBinaryProtocol> protocol(new TBinaryProtocol(transport));
         client = std::make_shared<TSIServiceClient>(protocol);
     }
 
@@ -552,10 +593,10 @@ void Session::close() {
     if (isClosed) {
         return;
     }
-    shared_ptr <TSCloseSessionReq> req(new TSCloseSessionReq());
+    shared_ptr<TSCloseSessionReq> req(new TSCloseSessionReq());
     req->__set_sessionId(sessionId);
     try {
-        shared_ptr <TSStatus> resp(new TSStatus());
+        shared_ptr<TSStatus> resp(new TSStatus());
         client->closeSession(*resp, *req);
     }
     catch (exception e) {
@@ -570,8 +611,8 @@ void Session::close() {
 
 
 void Session::insertRecord(const string &deviceId, int64_t time,
-                           const vector <string> &measurements,
-                           const vector <string> &values) {
+                           const vector<string> &measurements,
+                           const vector<string> &values) {
     TSInsertStringRecordReq req;
     req.__set_sessionId(sessionId);
     req.__set_prefixPath(deviceId);
@@ -590,8 +631,8 @@ void Session::insertRecord(const string &deviceId, int64_t time,
 }
 
 void Session::insertRecord(const string &prefixPath, int64_t time,
-                           const vector <string> &measurements,
-                           const vector <TSDataType::TSDataType> &types,
+                           const vector<string> &measurements,
+                           const vector<TSDataType::TSDataType> &types,
                            const vector<char *> &values) {
     TSInsertRecordReq req;
     req.__set_sessionId(sessionId);
@@ -612,8 +653,8 @@ void Session::insertRecord(const string &prefixPath, int64_t time,
 }
 
 void Session::insertAlignedRecord(const string &deviceId, int64_t time,
-                           const vector <string> &measurements,
-                           const vector <string> &values) {
+                                  const vector<string> &measurements,
+                                  const vector<string> &values) {
     TSInsertStringRecordReq req;
     req.__set_sessionId(sessionId);
     req.__set_prefixPath(deviceId);
@@ -632,9 +673,9 @@ void Session::insertAlignedRecord(const string &deviceId, int64_t time,
 }
 
 void Session::insertAlignedRecord(const string &prefixPath, int64_t time,
-                           const vector <string> &measurements,
-                           const vector <TSDataType::TSDataType> &types,
-                           const vector<char *> &values) {
+                                  const vector<string> &measurements,
+                                  const vector<TSDataType::TSDataType> &types,
+                                  const vector<char *> &values) {
     TSInsertRecordReq req;
     req.__set_sessionId(sessionId);
     req.__set_prefixPath(prefixPath);
@@ -653,10 +694,10 @@ void Session::insertAlignedRecord(const string &prefixPath, int64_t time,
     }
 }
 
-void Session::insertRecords(const vector <string> &deviceIds,
-                            const vector <int64_t> &times,
-                            const vector <vector<string>> &measurementsList,
-                            const vector <vector<string>> &valuesList) {
+void Session::insertRecords(const vector<string> &deviceIds,
+                            const vector<int64_t> &times,
+                            const vector<vector<string>> &measurementsList,
+                            const vector<vector<string>> &valuesList) {
     size_t len = deviceIds.size();
     if (len != times.size() || len != measurementsList.size() || len != valuesList.size()) {
         logic_error e("deviceIds, times, measurementsList and valuesList's size should be equal");
@@ -680,11 +721,11 @@ void Session::insertRecords(const vector <string> &deviceIds,
     }
 }
 
-void Session::insertRecords(const vector <string> &deviceIds,
-                            const vector <int64_t> &times,
-                            const vector <vector<string>> &measurementsList,
-                            const vector <vector<TSDataType::TSDataType>> &typesList,
-                            const vector <vector<char *>> &valuesList) {
+void Session::insertRecords(const vector<string> &deviceIds,
+                            const vector<int64_t> &times,
+                            const vector<vector<string>> &measurementsList,
+                            const vector<vector<TSDataType::TSDataType>> &typesList,
+                            const vector<vector<char *>> &valuesList) {
     size_t len = deviceIds.size();
     if (len != times.size() || len != measurementsList.size() || len != valuesList.size()) {
         logic_error e("deviceIds, times, measurementsList and valuesList's size should be equal");
@@ -695,7 +736,7 @@ void Session::insertRecords(const vector <string> &deviceIds,
     request.__set_prefixPaths(deviceIds);
     request.__set_timestamps(times);
     request.__set_measurementsList(measurementsList);
-    vector <string> bufferList;
+    vector<string> bufferList;
     for (size_t i = 0; i < valuesList.size(); i++) {
         string buffer;
         putValuesIntoBuffer(typesList[i], valuesList[i], buffer);
@@ -713,10 +754,10 @@ void Session::insertRecords(const vector <string> &deviceIds,
     }
 }
 
-void Session::insertAlignedRecords(const vector <string> &deviceIds,
-                            const vector <int64_t> &times,
-                            const vector <vector<string>> &measurementsList,
-                            const vector <vector<string>> &valuesList) {
+void Session::insertAlignedRecords(const vector<string> &deviceIds,
+                                   const vector<int64_t> &times,
+                                   const vector<vector<string>> &measurementsList,
+                                   const vector<vector<string>> &valuesList) {
     size_t len = deviceIds.size();
     if (len != times.size() || len != measurementsList.size() || len != valuesList.size()) {
         logic_error e("deviceIds, times, measurementsList and valuesList's size should be equal");
@@ -740,11 +781,11 @@ void Session::insertAlignedRecords(const vector <string> &deviceIds,
     }
 }
 
-void Session::insertAlignedRecords(const vector <string> &deviceIds,
-                            const vector <int64_t> &times,
-                            const vector <vector<string>> &measurementsList,
-                            const vector <vector<TSDataType::TSDataType>> &typesList,
-                            const vector <vector<char *>> &valuesList) {
+void Session::insertAlignedRecords(const vector<string> &deviceIds,
+                                   const vector<int64_t> &times,
+                                   const vector<vector<string>> &measurementsList,
+                                   const vector<vector<TSDataType::TSDataType>> &typesList,
+                                   const vector<vector<char *>> &valuesList) {
     size_t len = deviceIds.size();
     if (len != times.size() || len != measurementsList.size() || len != valuesList.size()) {
         logic_error e("deviceIds, times, measurementsList and valuesList's size should be equal");
@@ -755,7 +796,7 @@ void Session::insertAlignedRecords(const vector <string> &deviceIds,
     request.__set_prefixPaths(deviceIds);
     request.__set_timestamps(times);
     request.__set_measurementsList(measurementsList);
-    vector <string> bufferList;
+    vector<string> bufferList;
     for (size_t i = 0; i < valuesList.size(); i++) {
         string buffer;
         putValuesIntoBuffer(typesList[i], valuesList[i], buffer);
@@ -774,18 +815,18 @@ void Session::insertAlignedRecords(const vector <string> &deviceIds,
 }
 
 void Session::insertRecordsOfOneDevice(const string &deviceId,
-                                       vector <int64_t> &times,
-                                       vector <vector<string>> &measurementsList,
-                                       vector <vector<TSDataType::TSDataType>> &typesList,
-                                       vector <vector<char *>> &valuesList) {
+                                       vector<int64_t> &times,
+                                       vector<vector<string>> &measurementsList,
+                                       vector<vector<TSDataType::TSDataType>> &typesList,
+                                       vector<vector<char *>> &valuesList) {
     insertRecordsOfOneDevice(deviceId, times, measurementsList, typesList, valuesList, false);
 }
 
 void Session::insertRecordsOfOneDevice(const string &deviceId,
-                                       vector <int64_t> &times,
-                                       vector <vector<string>> &measurementsList,
-                                       vector <vector<TSDataType::TSDataType>> &typesList,
-                                       vector <vector<char *>> &valuesList,
+                                       vector<int64_t> &times,
+                                       vector<vector<string>> &measurementsList,
+                                       vector<vector<TSDataType::TSDataType>> &typesList,
+                                       vector<vector<char *>> &valuesList,
                                        bool sorted) {
 
     if (sorted) {
@@ -810,7 +851,7 @@ void Session::insertRecordsOfOneDevice(const string &deviceId,
     request.__set_prefixPath(deviceId);
     request.__set_timestamps(times);
     request.__set_measurementsList(measurementsList);
-    vector <string> bufferList;
+    vector<string> bufferList;
     for (size_t i = 0; i < valuesList.size(); i++) {
         string buffer;
         putValuesIntoBuffer(typesList[i], valuesList[i], buffer);
@@ -829,19 +870,19 @@ void Session::insertRecordsOfOneDevice(const string &deviceId,
 }
 
 void Session::insertAlignedRecordsOfOneDevice(const string &deviceId,
-                                       vector <int64_t> &times,
-                                       vector <vector<string>> &measurementsList,
-                                       vector <vector<TSDataType::TSDataType>> &typesList,
-                                       vector <vector<char *>> &valuesList) {
+                                              vector<int64_t> &times,
+                                              vector<vector<string>> &measurementsList,
+                                              vector<vector<TSDataType::TSDataType>> &typesList,
+                                              vector<vector<char *>> &valuesList) {
     insertAlignedRecordsOfOneDevice(deviceId, times, measurementsList, typesList, valuesList, false);
 }
 
 void Session::insertAlignedRecordsOfOneDevice(const string &deviceId,
-                                       vector <int64_t> &times,
-                                       vector <vector<string>> &measurementsList,
-                                       vector <vector<TSDataType::TSDataType>> &typesList,
-                                       vector <vector<char *>> &valuesList,
-                                       bool sorted) {
+                                              vector<int64_t> &times,
+                                              vector<vector<string>> &measurementsList,
+                                              vector<vector<TSDataType::TSDataType>> &typesList,
+                                              vector<vector<char *>> &valuesList,
+                                              bool sorted) {
 
     if (sorted) {
         if (!checkSorted(times)) {
@@ -865,7 +906,7 @@ void Session::insertAlignedRecordsOfOneDevice(const string &deviceId,
     request.__set_prefixPath(deviceId);
     request.__set_timestamps(times);
     request.__set_measurementsList(measurementsList);
-    vector <string> bufferList;
+    vector<string> bufferList;
     for (size_t i = 0; i < valuesList.size(); i++) {
         string buffer;
         putValuesIntoBuffer(typesList[i], valuesList[i], buffer);
@@ -905,7 +946,7 @@ void Session::insertTablet(Tablet &tablet, bool sorted) {
     TSInsertTabletReq request;
     request.__set_sessionId(sessionId);
     request.prefixPath = tablet.deviceId;
-    for (pair <string, TSDataType::TSDataType> schema: tablet.schemas) {
+    for (pair<string, TSDataType::TSDataType> schema: tablet.schemas) {
         request.measurements.push_back(schema.first);
         request.types.push_back(schema.second);
     }
@@ -969,9 +1010,9 @@ void Session::insertTablets(map<string, Tablet *> &tablets, bool sorted) {
             sortTablet(*(item.second));
         }
         request.prefixPaths.push_back(item.second->deviceId);
-        vector <string> measurements;
+        vector<string> measurements;
         vector<int> dataTypes;
-        for (pair <string, TSDataType::TSDataType> schema: item.second->schemas) {
+        for (pair<string, TSDataType::TSDataType> schema: item.second->schemas) {
             measurements.push_back(schema.first);
             dataTypes.push_back(schema.second);
         }
@@ -997,7 +1038,7 @@ void Session::insertAlignedTablets(map<string, Tablet *> &tablets) {
 }
 
 void Session::insertAlignedTablets(map<string, Tablet *> &tablets, bool sorted) {
-    for(map<string, Tablet *>::iterator iter = tablets.begin(); iter != tablets.end(); iter++) {
+    for (map<string, Tablet *>::iterator iter = tablets.begin(); iter != tablets.end(); iter++) {
         iter->second->setAligned(true);
     }
     try {
@@ -1009,15 +1050,15 @@ void Session::insertAlignedTablets(map<string, Tablet *> &tablets, bool sorted) 
     }
 }
 
-void Session::testInsertRecord(const string &deviceId, int64_t time, const vector <string> &measurements,
-                               const vector <string> &values) {
-    shared_ptr <TSInsertStringRecordReq> req(new TSInsertStringRecordReq());
+void Session::testInsertRecord(const string &deviceId, int64_t time, const vector<string> &measurements,
+                               const vector<string> &values) {
+    shared_ptr<TSInsertStringRecordReq> req(new TSInsertStringRecordReq());
     req->__set_sessionId(sessionId);
     req->__set_prefixPath(deviceId);
     req->__set_timestamp(time);
     req->__set_measurements(measurements);
     req->__set_values(values);
-    shared_ptr <TSStatus> resp(new TSStatus());
+    shared_ptr<TSStatus> resp(new TSStatus());
     try {
         client->insertStringRecord(*resp, *req);
         RpcUtils::verifySuccess(*resp);
@@ -1028,10 +1069,10 @@ void Session::testInsertRecord(const string &deviceId, int64_t time, const vecto
 }
 
 void Session::testInsertTablet(const Tablet &tablet) {
-    shared_ptr <TSInsertTabletReq> request(new TSInsertTabletReq());
+    shared_ptr<TSInsertTabletReq> request(new TSInsertTabletReq());
     request->__set_sessionId(sessionId);
     request->prefixPath = tablet.deviceId;
-    for (pair <string, TSDataType::TSDataType> schema: tablet.schemas) {
+    for (pair<string, TSDataType::TSDataType> schema: tablet.schemas) {
         request->measurements.push_back(schema.first);
         request->types.push_back(schema.second);
     }
@@ -1040,7 +1081,7 @@ void Session::testInsertTablet(const Tablet &tablet) {
     request->__set_size(tablet.rowSize);
 
     try {
-        shared_ptr <TSStatus> resp(new TSStatus());
+        shared_ptr<TSStatus> resp(new TSStatus());
         client->testInsertTablet(*resp, *request);
         RpcUtils::verifySuccess(*resp);
     }
@@ -1049,16 +1090,16 @@ void Session::testInsertTablet(const Tablet &tablet) {
     }
 }
 
-void Session::testInsertRecords(const vector <string> &deviceIds,
-                                const vector <int64_t> &times,
-                                const vector <vector<string>> &measurementsList,
-                                const vector <vector<string>> &valuesList) {
+void Session::testInsertRecords(const vector<string> &deviceIds,
+                                const vector<int64_t> &times,
+                                const vector<vector<string>> &measurementsList,
+                                const vector<vector<string>> &valuesList) {
     size_t len = deviceIds.size();
     if (len != times.size() || len != measurementsList.size() || len != valuesList.size()) {
         logic_error error("deviceIds, times, measurementsList and valuesList's size should be equal");
         throw exception(error);
     }
-    shared_ptr <TSInsertStringRecordsReq> request(new TSInsertStringRecordsReq());
+    shared_ptr<TSInsertStringRecordsReq> request(new TSInsertStringRecordsReq());
     request->__set_sessionId(sessionId);
     request->__set_prefixPaths(deviceIds);
     request->__set_timestamps(times);
@@ -1066,7 +1107,7 @@ void Session::testInsertRecords(const vector <string> &deviceIds,
     request->__set_valuesList(valuesList);
 
     try {
-        shared_ptr <TSStatus> resp(new TSStatus());
+        shared_ptr<TSStatus> resp(new TSStatus());
         client->insertStringRecords(*resp, *request);
         RpcUtils::verifySuccess(*resp);
     }
@@ -1076,13 +1117,13 @@ void Session::testInsertRecords(const vector <string> &deviceIds,
 }
 
 void Session::deleteTimeseries(const string &path) {
-    vector <string> paths;
+    vector<string> paths;
     paths.push_back(path);
     deleteTimeseries(paths);
 }
 
-void Session::deleteTimeseries(const vector <string> &paths) {
-    shared_ptr <TSStatus> resp(new TSStatus());
+void Session::deleteTimeseries(const vector<string> &paths) {
+    shared_ptr<TSStatus> resp(new TSStatus());
     try {
         client->deleteTimeseries(*resp, sessionId, paths);
         RpcUtils::verifySuccess(*resp);
@@ -1093,17 +1134,17 @@ void Session::deleteTimeseries(const vector <string> &paths) {
 }
 
 void Session::deleteData(const string &path, int64_t time) {
-    vector <string> paths;
+    vector<string> paths;
     paths.push_back(path);
     deleteData(paths, time);
 }
 
-void Session::deleteData(const vector <string> &deviceId, int64_t time) {
-    shared_ptr <TSDeleteDataReq> req(new TSDeleteDataReq());
+void Session::deleteData(const vector<string> &deviceId, int64_t time) {
+    shared_ptr<TSDeleteDataReq> req(new TSDeleteDataReq());
     req->__set_sessionId(sessionId);
     req->__set_paths(deviceId);
     req->__set_endTime(time);
-    shared_ptr <TSStatus> resp(new TSStatus());
+    shared_ptr<TSStatus> resp(new TSStatus());
     try {
         client->deleteData(*resp, *req);
         RpcUtils::verifySuccess(*resp);
@@ -1114,7 +1155,7 @@ void Session::deleteData(const vector <string> &deviceId, int64_t time) {
 }
 
 void Session::setStorageGroup(const string &storageGroupId) {
-    shared_ptr <TSStatus> resp(new TSStatus());
+    shared_ptr<TSStatus> resp(new TSStatus());
     try {
         client->setStorageGroup(*resp, sessionId, storageGroupId);
         RpcUtils::verifySuccess(*resp);
@@ -1125,13 +1166,13 @@ void Session::setStorageGroup(const string &storageGroupId) {
 }
 
 void Session::deleteStorageGroup(const string &storageGroup) {
-    vector <string> storageGroups;
+    vector<string> storageGroups;
     storageGroups.push_back(storageGroup);
     deleteStorageGroups(storageGroups);
 }
 
-void Session::deleteStorageGroups(const vector <string> &storageGroups) {
-    shared_ptr <TSStatus> resp(new TSStatus());
+void Session::deleteStorageGroups(const vector<string> &storageGroups) {
+    shared_ptr<TSStatus> resp(new TSStatus());
     try {
         client->deleteStorageGroups(*resp, sessionId, storageGroups);
         RpcUtils::verifySuccess(*resp);
@@ -1158,11 +1199,11 @@ void Session::createTimeseries(const string &path,
                                TSDataType::TSDataType dataType,
                                TSEncoding::TSEncoding encoding,
                                CompressionType::CompressionType compressor,
-                               map <string, string> *props,
-                               map <string, string> *tags,
-                               map <string, string> *attributes,
+                               map<string, string> *props,
+                               map<string, string> *tags,
+                               map<string, string> *attributes,
                                const string &measurementAlias) {
-    shared_ptr <TSCreateTimeseriesReq> req(new TSCreateTimeseriesReq());
+    shared_ptr<TSCreateTimeseriesReq> req(new TSCreateTimeseriesReq());
     req->__set_sessionId(sessionId);
     req->__set_path(path);
     req->__set_dataType(dataType);
@@ -1182,7 +1223,7 @@ void Session::createTimeseries(const string &path,
         req->__set_measurementAlias(measurementAlias);
     }
 
-    shared_ptr <TSStatus> resp(new TSStatus());
+    shared_ptr<TSStatus> resp(new TSStatus());
     try {
         client->createTimeseries(*resp, *req);
         RpcUtils::verifySuccess(*resp);
@@ -1192,15 +1233,15 @@ void Session::createTimeseries(const string &path,
     }
 }
 
-void Session::createMultiTimeseries(const vector <string> &paths,
-                                    const vector <TSDataType::TSDataType> &dataTypes,
-                                    const vector <TSEncoding::TSEncoding> &encodings,
-                                    const vector <CompressionType::CompressionType> &compressors,
-                                    vector <map<string, string>> *propsList,
-                                    vector <map<string, string>> *tagsList,
-                                    vector <map<string, string>> *attributesList,
-                                    vector <string> *measurementAliasList) {
-    shared_ptr <TSCreateMultiTimeseriesReq> request(new TSCreateMultiTimeseriesReq());
+void Session::createMultiTimeseries(const vector<string> &paths,
+                                    const vector<TSDataType::TSDataType> &dataTypes,
+                                    const vector<TSEncoding::TSEncoding> &encodings,
+                                    const vector<CompressionType::CompressionType> &compressors,
+                                    vector<map<string, string>> *propsList,
+                                    vector<map<string, string>> *tagsList,
+                                    vector<map<string, string>> *attributesList,
+                                    vector<string> *measurementAliasList) {
+    shared_ptr<TSCreateMultiTimeseriesReq> request(new TSCreateMultiTimeseriesReq());
     request->__set_sessionId(sessionId);
     request->__set_paths(paths);
 
@@ -1237,7 +1278,7 @@ void Session::createMultiTimeseries(const vector <string> &paths,
     }
 
     try {
-        shared_ptr <TSStatus> resp(new TSStatus());
+        shared_ptr<TSStatus> resp(new TSStatus());
         client->createMultiTimeseries(*resp, *request);
         RpcUtils::verifySuccess(*resp);
     }
@@ -1247,11 +1288,11 @@ void Session::createMultiTimeseries(const vector <string> &paths,
 }
 
 void Session::createAlignedTimeseries(const std::string &deviceId,
-                             const std::vector <std::string> &measurements,
-                             const std::vector <TSDataType::TSDataType> &dataTypes,
-                             const std::vector <TSEncoding::TSEncoding> &encodings,
-                             const std::vector <CompressionType::CompressionType> &compressors) {
-    shared_ptr <TSCreateAlignedTimeseriesReq> request(new TSCreateAlignedTimeseriesReq());
+                                      const std::vector<std::string> &measurements,
+                                      const std::vector<TSDataType::TSDataType> &dataTypes,
+                                      const std::vector<TSEncoding::TSEncoding> &encodings,
+                                      const std::vector<CompressionType::CompressionType> &compressors) {
+    shared_ptr<TSCreateAlignedTimeseriesReq> request(new TSCreateAlignedTimeseriesReq());
     request->__set_sessionId(sessionId);
     request->__set_prefixPath(deviceId);
     request->__set_measurements(measurements);
@@ -1275,7 +1316,7 @@ void Session::createAlignedTimeseries(const std::string &deviceId,
     request->__set_compressors(compressorsOrdinal);
 
     try {
-        shared_ptr <TSStatus> resp(new TSStatus());
+        shared_ptr<TSStatus> resp(new TSStatus());
         client->createAlignedTimeseries(*resp, *request);
         RpcUtils::verifySuccess(*resp);
     }
@@ -1286,12 +1327,13 @@ void Session::createAlignedTimeseries(const std::string &deviceId,
 
 bool Session::checkTimeseriesExists(const string &path) {
     try {
-        std::unique_ptr <SessionDataSet> dataset = executeQueryStatement("SHOW TIMESERIES " + path);
+        std::unique_ptr<SessionDataSet> dataset = executeQueryStatement("SHOW TIMESERIES " + path);
         bool isExisted = dataset->hasNext();
         dataset->closeOperationHandle();
         return isExisted;
     }
-    catch (exception e) {
+    catch (exception &e) {
+        std::cout << e.what() << std::endl;
         throw IoTDBConnectionException(e.what());
     }
 }
@@ -1300,7 +1342,7 @@ string Session::getTimeZone() {
     if (zoneId != "") {
         return zoneId;
     }
-    shared_ptr <TSGetTimeZoneResp> resp(new TSGetTimeZoneResp());
+    shared_ptr<TSGetTimeZoneResp> resp(new TSGetTimeZoneResp());
     try {
         client->getTimeZone(*resp, sessionId);
         RpcUtils::verifySuccess(resp->status);
@@ -1312,10 +1354,10 @@ string Session::getTimeZone() {
 }
 
 void Session::setTimeZone(const string &zoneId) {
-    shared_ptr <TSSetTimeZoneReq> req(new TSSetTimeZoneReq());
+    shared_ptr<TSSetTimeZoneReq> req(new TSSetTimeZoneReq());
     req->__set_sessionId(sessionId);
     req->__set_timeZone(zoneId);
-    shared_ptr <TSStatus> resp(new TSStatus());
+    shared_ptr<TSStatus> resp(new TSStatus());
     try {
         client->setTimeZone(*resp, *req);
     }
@@ -1326,13 +1368,13 @@ void Session::setTimeZone(const string &zoneId) {
     this->zoneId = zoneId;
 }
 
-unique_ptr <SessionDataSet> Session::executeQueryStatement(const string &sql) {
-    shared_ptr <TSExecuteStatementReq> req(new TSExecuteStatementReq());
+unique_ptr<SessionDataSet> Session::executeQueryStatement(const string &sql) {
+    shared_ptr<TSExecuteStatementReq> req(new TSExecuteStatementReq());
     req->__set_sessionId(sessionId);
     req->__set_statementId(statementId);
     req->__set_statement(sql);
     req->__set_fetchSize(fetchSize);
-    shared_ptr <TSExecuteStatementResp> resp(new TSExecuteStatementResp());
+    shared_ptr<TSExecuteStatementResp> resp(new TSExecuteStatementResp());
     try {
         client->executeStatement(*resp, *req);
         RpcUtils::verifySuccess(resp->status);
@@ -1340,17 +1382,18 @@ unique_ptr <SessionDataSet> Session::executeQueryStatement(const string &sql) {
     catch (IoTDBConnectionException e) {
         throw IoTDBConnectionException(e.what());
     }
-    shared_ptr <TSQueryDataSet> queryDataSet(new TSQueryDataSet(resp->queryDataSet));
+    shared_ptr<TSQueryDataSet> queryDataSet(new TSQueryDataSet(resp->queryDataSet));
     return unique_ptr<SessionDataSet>(new SessionDataSet(
-            sql, resp->columns, resp->dataTypeList, resp->queryId, statementId, client, sessionId, queryDataSet));
+            sql, resp->columns, resp->dataTypeList, resp->columnNameIndexMap, resp->ignoreTimeStamp, resp->queryId,
+            statementId, client, sessionId, queryDataSet));
 }
 
 void Session::executeNonQueryStatement(const string &sql) {
-    shared_ptr <TSExecuteStatementReq> req(new TSExecuteStatementReq());
+    shared_ptr<TSExecuteStatementReq> req(new TSExecuteStatementReq());
     req->__set_sessionId(sessionId);
     req->__set_statementId(statementId);
     req->__set_statement(sql);
-    shared_ptr <TSExecuteStatementResp> resp(new TSExecuteStatementResp());
+    shared_ptr<TSExecuteStatementResp> resp(new TSExecuteStatementResp());
     try {
         client->executeUpdateStatement(*resp, *req);
         RpcUtils::verifySuccess(resp->status);
