@@ -26,12 +26,12 @@ Session *session;
 void createAlignedTimeseries() {
     string alignedDeviceId = "root.sg1.d1";
     vector<string> measurements = {"s1", "s2", "s3"};
-    vector<string> alignedTimeseries = {"root.sg1.d1.s1","root.sg1.d1.s2","root.sg1.d1.s3"};
+    vector<string> alignedTimeseries = {"root.sg1.d1.s1", "root.sg1.d1.s2", "root.sg1.d1.s3"};
     vector<TSDataType::TSDataType> dataTypes = {TSDataType::INT32, TSDataType::DOUBLE, TSDataType::BOOLEAN};
     vector<TSEncoding::TSEncoding> encodings = {TSEncoding::PLAIN, TSEncoding::GORILLA, TSEncoding::RLE};
     vector<CompressionType::CompressionType> compressors = {
-        CompressionType::SNAPPY, CompressionType::UNCOMPRESSED, CompressionType::SNAPPY};
-    for (string timeseries : alignedTimeseries) {
+            CompressionType::SNAPPY, CompressionType::UNCOMPRESSED, CompressionType::SNAPPY};
+    for (const string &timeseries: alignedTimeseries) {
         if (session->checkTimeseriesExists(timeseries)) {
             session->deleteTimeseries(timeseries);
         }
@@ -39,18 +39,34 @@ void createAlignedTimeseries() {
     session->createAlignedTimeseries(alignedDeviceId, measurements, dataTypes, encodings, compressors);
 }
 
+void showTimeseries() {
+    unique_ptr<SessionDataSet> dataSet = session->executeQueryStatement("show timeseries");
+    for (const string &name: dataSet->getColumnNames()) {
+        cout << name << "  ";
+    }
+    cout << endl;
+
+    dataSet->setBatchSize(1024);
+    while (dataSet->hasNext()) {
+        cout << dataSet->next()->toString();
+    }
+    cout << endl;
+
+    dataSet->closeOperationHandle();
+}
+
 void insertAlignedRecord() {
     string deviceId = "root.sg1.d1";
     vector<string> measurements;
-    measurements.push_back("s1");
-    measurements.push_back("s2");
-    measurements.push_back("s3");
+    measurements.emplace_back("s1");
+    measurements.emplace_back("s2");
+    measurements.emplace_back("s3");
 
-    for (int64_t time = 0; time < 100; time++) {
+    for (int64_t time = 0; time < 10; time++) {
         vector<string> values;
-        values.push_back("1");
-        values.push_back("1.0");
-        values.push_back("true");
+        values.emplace_back("1");
+        values.emplace_back("1.0");
+        values.emplace_back("true");
         session->insertAlignedRecord(deviceId, time, measurements, values);
     }
 }
@@ -58,26 +74,26 @@ void insertAlignedRecord() {
 void insertAlignedRecords() {
     string deviceId = "root.sg1.d1";
     vector<string> measurements;
-    measurements.push_back("s1");
-    measurements.push_back("s2");
-    measurements.push_back("s3");
+    measurements.emplace_back("s1");
+    measurements.emplace_back("s2");
+    measurements.emplace_back("s3");
 
     vector<string> deviceIds;
     vector<vector<string>> measurementsList;
     vector<vector<string>> valuesList;
     vector<int64_t> timestamps;
 
-    for (int64_t time = 100; time < 200; time++) {
+    for (int64_t time = 10; time < 20; time++) {
         vector<string> values;
-        values.push_back("1");
-        values.push_back("1.0");
-        values.push_back("true");
+        values.emplace_back("1");
+        values.emplace_back("1.0");
+        values.emplace_back("true");
 
         deviceIds.push_back(deviceId);
         measurementsList.push_back(measurements);
         valuesList.push_back(values);
         timestamps.push_back(time);
-        if (time != 100 && time % 100 == 0) {
+        if (time != 10 && time % 10 == 0) {
             session->insertAlignedRecords(deviceIds, timestamps, measurementsList, valuesList);
             deviceIds.clear();
             measurementsList.clear();
@@ -98,10 +114,10 @@ void insertAlignedTablet() {
     schemas.push_back(pairB);
     schemas.push_back(pairC);
 
-    Tablet tablet("root.sg1.d1", schemas, 100);
+    Tablet tablet("root.sg1.d1", schemas, 10);
     tablet.setAligned(true);
 
-    for (int64_t time = 200; time < 300; time++) {
+    for (int64_t time = 20; time < 30; time++) {
         int row = tablet.rowSize++;
         tablet.timestamps[row] = time;
         tablet.values[0][row] = "1";
@@ -128,16 +144,16 @@ void insertAlignedTablets() {
     schemas.push_back(pairB);
     schemas.push_back(pairC);
 
-    Tablet tablet1("root.sg1.d1", schemas, 100);
-    Tablet tablet2("root.sg1.d2", schemas, 100);
-    Tablet tablet3("root.sg1.d3", schemas, 100);
+    Tablet tablet1("root.sg1.d1", schemas, 10);
+    Tablet tablet2("root.sg1.d2", schemas, 10);
+    Tablet tablet3("root.sg1.d3", schemas, 10);
 
-    map<string, Tablet*> tabletMap;
+    map<string, Tablet *> tabletMap;
     tabletMap["root.sg1.d1"] = &tablet1;
     tabletMap["root.sg1.d2"] = &tablet2;
     tabletMap["root.sg1.d3"] = &tablet3;
 
-    for (int64_t time = 300; time < 400; time++) {
+    for (int64_t time = 30; time < 40; time++) {
         int row1 = tablet1.rowSize++;
         int row2 = tablet2.rowSize++;
         int row3 = tablet3.rowSize++;
@@ -174,6 +190,80 @@ void insertAlignedTablets() {
     }
 }
 
+void insertNullableTabletWithAlignedTimeseries() {
+    pair<string, TSDataType::TSDataType> pairA("s1", TSDataType::INT32);
+    pair<string, TSDataType::TSDataType> pairB("s2", TSDataType::DOUBLE);
+    pair<string, TSDataType::TSDataType> pairC("s3", TSDataType::BOOLEAN);
+    vector<pair<string, TSDataType::TSDataType>> schemas;
+    schemas.push_back(pairA);
+    schemas.push_back(pairB);
+    schemas.push_back(pairC);
+
+    Tablet tablet("root.sg1.d1", schemas, 10);
+    tablet.setAligned(true);
+
+    for (int64_t time = 40; time < 50; time++) {
+        int row = tablet.rowSize++;
+        tablet.timestamps[row] = time;
+        for (int i = 0; i < 3; i++) {
+            if (i == 0) {
+                tablet.values[i][row] = "1";
+            } else if (i == 1) {
+                tablet.values[i][row] = "1.0";
+            } else {
+                tablet.values[i][row] = "true";
+            }
+            // mark null value
+            if (row % 3 == i) {
+                tablet.bitMaps[i]->mark(row);
+            }
+        }
+        if (tablet.rowSize == tablet.maxRowNumber) {
+            session->insertTablet(tablet, true);
+            tablet.reset();
+        }
+    }
+
+    if (tablet.rowSize != 0) {
+        session->insertTablet(tablet);
+        tablet.reset();
+    }
+}
+
+void query() {
+    unique_ptr<SessionDataSet> dataSet = session->executeQueryStatement("select * from root.sg1.d1");
+    cout << "timestamp" << "  ";
+    for (const string &name: dataSet->getColumnNames()) {
+        cout << name << "  ";
+    }
+    cout << endl;
+
+    dataSet->setBatchSize(1024);
+    while (dataSet->hasNext()) {
+        cout << dataSet->next()->toString();
+    }
+    cout << endl;
+
+    dataSet->closeOperationHandle();
+}
+
+void deleteData() {
+    string path = "root.sg1.d1.s1";
+    int64_t deleteTime = 49;
+    session->deleteData(path, deleteTime);
+}
+
+void deleteTimeseries() {
+    vector<string> paths;
+    vector<string> alignedTimeseries = {"root.sg1.d1.s1", "root.sg1.d1.s2", "root.sg1.d1.s3"};
+    for (const string &timeseries: alignedTimeseries) {
+        if (session->checkTimeseriesExists(timeseries)) {
+            paths.push_back(timeseries);
+        }
+    }
+    session->deleteTimeseries(paths);
+}
+
 int main() {
     session = new Session("127.0.0.1", 6667, "root", "root");
 
@@ -184,7 +274,7 @@ int main() {
     try {
         session->setStorageGroup("root.sg1");
     }
-    catch (IoTDBConnectionException e){
+    catch (IoTDBConnectionException &e) {
         string errorMessage(e.what());
         if (errorMessage.find("StorageGroupAlreadySetException") == string::npos) {
             cout << errorMessage << endl;
@@ -194,6 +284,9 @@ int main() {
 
     cout << "createAlignedTimeseries\n" << endl;
     createAlignedTimeseries();
+
+    cout << "showTimeseries\n" << endl;
+    showTimeseries();
 
     cout << "insertAlignedRecord\n" << endl;
     insertAlignedRecord();
@@ -206,6 +299,18 @@ int main() {
 
     cout << "insertAlignedTablets\n" << endl;
     insertAlignedTablets();
+
+    cout << "insertNullableTabletWithAlignedTimeseries\n" << endl;
+    insertNullableTabletWithAlignedTimeseries();
+
+    cout << "query\n" << endl;
+    query();
+
+    cout << "deleteData\n" << endl;
+    deleteData();
+
+    cout << "deleteTimeseries\n" << endl;
+    deleteTimeseries();
 
     cout << "session close\n" << endl;
     session->close();
