@@ -59,6 +59,8 @@ public class IoTDBSelectIntoIT {
     "insert into root.sg.d1(time, s1, s2, s4, s5, s6) values (2, 2, 2, 2, true, '2')",
     "insert into root.sg.d1(time, s1, s2, s3, s5, s6) values (3, 3, 3, 3, false, '3')",
     "insert into root.sg.d1(time, s1, s2, s3, s4, s5, s6) values (4, 4, 4, 4, 4, true, '4')",
+    "insert into root.sg.d1(time, s1, s2, s3, s4, s6) values (5, 5, 5, 5, 5, '5')",
+    "insert into root.sg.d1(time, s1, s2, s3, s4, s5) values (6, 6, 6, 6, 6, true)",
   };
 
   @BeforeClass
@@ -361,6 +363,61 @@ public class IoTDBSelectIntoIT {
   }
 
   @Test
+  public void testSelectIntoAlignedTimeSeriesCorrectly() {
+    try (Connection connection =
+            DriverManager.getConnection(
+                Config.IOTDB_URL_PREFIX + "127.0.0.1:6667/", "root", "root");
+        Statement statement = connection.createStatement()) {
+      statement.execute(
+          "select s1, s1 "
+              + "into aligned root.sg.`aligned`.s1s2, root.sg.`aligned`.s1s3 "
+              + "from root.sg.d1 "
+              + "where time <= 2");
+      statement.execute(
+          "select s1, s1 "
+              + "into aligned root.sg.`aligned`.s1s2, root.sg.`aligned`.s1s3 "
+              + "from root.sg.d1 "
+              + "where time > 2");
+
+      try (ResultSet resultSet =
+          statement.executeQuery("select s1s2, s1s3 from root.sg.`aligned`")) {
+        assertEquals(1 + 2, resultSet.getMetaData().getColumnCount());
+
+        for (int i = 1; i < INSERTION_SQLS.length; ++i) {
+          assertTrue(resultSet.next());
+          for (int j = 0; j < 2 + 1; ++j) {
+            assertEquals(resultSet.getString(2), resultSet.getString(3));
+          }
+        }
+
+        assertFalse(resultSet.next());
+      }
+    } catch (SQLException throwable) {
+      fail(throwable.getMessage());
+    }
+  }
+
+  @Test
+  public void testSelectIntoAlignedTimeSeriesWithUnmatchedTypes() {
+    try (Connection connection =
+            DriverManager.getConnection(
+                Config.IOTDB_URL_PREFIX + "127.0.0.1:6667/", "root", "root");
+        Statement statement = connection.createStatement()) {
+      statement.execute("create aligned timeseries root.sg.`aligned`(s1 TEXT, s2 TEXT)");
+      statement.execute(
+          "select s1, s1 "
+              + "into aligned root.sg.`aligned`.s1, root.sg.`aligned`.s2 "
+              + "from root.sg.d1");
+      fail();
+    } catch (SQLException throwable) {
+      assertTrue(
+          throwable
+              .getMessage()
+              .contains("failed to insert measurements [s1, s2] caused by DataType mismatch"));
+    }
+  }
+
+  @Test
   public void testGroupByQuery() {
     try (Connection connection =
             DriverManager.getConnection(
@@ -371,7 +428,7 @@ public class IoTDBSelectIntoIT {
       try (ResultSet resultSet = statement.executeQuery("select count_s1 from root.sg.d1")) {
         assertEquals(1 + 1, resultSet.getMetaData().getColumnCount());
 
-        for (int i = 1; i < INSERTION_SQLS.length; ++i) {
+        for (int i = 1; i < INSERTION_SQLS.length - 2; ++i) {
           assertTrue(resultSet.next());
           for (int j = 0; j < 1 + 1; ++j) {
             assertEquals(String.valueOf(i), resultSet.getString(1));
@@ -406,6 +463,8 @@ public class IoTDBSelectIntoIT {
           }
         }
 
+        assertTrue(resultSet.next());
+        assertTrue(resultSet.next());
         assertFalse(resultSet.next());
       }
     } catch (SQLException throwable) {
@@ -427,7 +486,7 @@ public class IoTDBSelectIntoIT {
 
         assertTrue(resultSet.next());
         assertEquals("10", resultSet.getString(1));
-        assertEquals("4", resultSet.getString(2));
+        assertEquals("6", resultSet.getString(2));
 
         assertFalse(resultSet.next());
       }
