@@ -768,7 +768,9 @@ public class IoTDBAggregationIT {
       } catch (Exception e) {
         Assert.assertTrue(
             e.getMessage(),
-            e.getMessage().contains("Unsupported data type in aggregation AVG : TEXT"));
+            e.getMessage()
+                .contains(
+                    "Aggregate functions [AVG, SUM, EXTREME, MIN_VALUE, MAX_VALUE] only support numeric data types [INT32, INT64, FLOAT, DOUBLE]"));
       }
       try {
         statement.execute(
@@ -779,7 +781,9 @@ public class IoTDBAggregationIT {
         }
       } catch (Exception e) {
         Assert.assertTrue(
-            e.getMessage().contains("Unsupported data type in aggregation SUM : TEXT"));
+            e.getMessage()
+                .contains(
+                    "Aggregate functions [AVG, SUM, EXTREME, MIN_VALUE, MAX_VALUE] only support numeric data types [INT32, INT64, FLOAT, DOUBLE]"));
       }
       try {
         statement.execute(
@@ -790,7 +794,9 @@ public class IoTDBAggregationIT {
         }
       } catch (Exception e) {
         Assert.assertTrue(
-            e.getMessage().contains("Unsupported data type in aggregation AVG : BOOLEAN"));
+            e.getMessage()
+                .contains(
+                    "Aggregate functions [AVG, SUM, EXTREME, MIN_VALUE, MAX_VALUE] only support numeric data types [INT32, INT64, FLOAT, DOUBLE]"));
       }
       try {
         statement.execute(
@@ -802,7 +808,9 @@ public class IoTDBAggregationIT {
       } catch (Exception e) {
         Assert.assertTrue(
             e.getMessage(),
-            e.getMessage().contains("Unsupported data type in aggregation SUM : BOOLEAN"));
+            e.getMessage()
+                .contains(
+                    "Aggregate functions [AVG, SUM, EXTREME, MIN_VALUE, MAX_VALUE] only support numeric data types [INT32, INT64, FLOAT, DOUBLE]"));
       }
       try {
         statement.execute("SELECT avg(status) FROM root.ln.wf01.wt01");
@@ -812,7 +820,10 @@ public class IoTDBAggregationIT {
         }
       } catch (Exception e) {
         Assert.assertTrue(
-            e.getMessage(), e.getMessage().contains("Boolean statistics does not support: avg"));
+            e.getMessage(),
+            e.getMessage()
+                .contains(
+                    "Aggregate functions [AVG, SUM, EXTREME, MIN_VALUE, MAX_VALUE] only support numeric data types [INT32, INT64, FLOAT, DOUBLE]"));
       }
     } catch (Exception e) {
       e.printStackTrace();
