@@ -27,7 +27,7 @@ import org.apache.iotdb.db.qp.physical.crud.InsertRowPlan;
 import org.apache.iotdb.db.qp.physical.crud.InsertRowsPlan;
 import org.apache.iotdb.db.qp.strategy.PhysicalGenerator;
 
-import java.util.Arrays;
+import java.util.List;
 
 /** this class extends {@code RootOperator} and process insert statement. */
 public class InsertOperator extends Operator {
@@ -36,7 +36,7 @@ public class InsertOperator extends Operator {
 
   private long[] times;
   private String[] measurementList;
-  private String[] valueList;
+  private List<String[]> valueLists;
 
   private boolean isAligned;
 
@@ -61,12 +61,12 @@ public class InsertOperator extends Operator {
     this.measurementList = measurementList;
   }
 
-  public String[] getValueList() {
-    return valueList;
+  public List<String[]> getValueLists() {
+    return valueLists;
   }
 
-  public void setValueList(String[] insertValue) {
-    this.valueList = insertValue;
+  public void setValueLists(List<String[]> valueLists) {
+    this.valueLists = valueLists;
   }
 
   public long[] getTimes() {
@@ -88,33 +88,29 @@ public class InsertOperator extends Operator {
   @Override
   public PhysicalPlan generatePhysicalPlan(PhysicalGenerator generator)
       throws QueryProcessException {
-    int measurementsNum = 0;
-    for (String measurement : measurementList) {
-      if (measurement.startsWith("(") && measurement.endsWith(")")) {
-        measurementsNum += measurement.replace("(", "").replace(")", "").split(",").length;
-      } else {
-        measurementsNum++;
+    int measurementsNum = measurementList.length;
+    if (times.length == 1) {
+      if (measurementsNum != valueLists.get(0).length) {
+        throw new SQLParserException(
+            String.format(
+                "the measurementList's size %d is not consistent with the valueList's size %d",
+                measurementsNum, valueLists.get(0).length));
       }
-    }
-    if (measurementsNum == 0 || (valueList.length % measurementsNum != 0)) {
-      throw new SQLParserException(
-          String.format(
-              "the measurementList's size %d is not consistent with the valueList's size %d",
-              measurementsNum, valueList.length));
-    }
-    if (measurementsNum == valueList.length) {
-      InsertRowPlan insertRowPlan = new InsertRowPlan(device, times[0], measurementList, valueList);
+      InsertRowPlan insertRowPlan =
+          new InsertRowPlan(device, times[0], measurementList, valueLists.get(0));
       insertRowPlan.setAligned(isAligned);
       return insertRowPlan;
     }
     InsertRowsPlan insertRowsPlan = new InsertRowsPlan();
     for (int i = 0; i < times.length; i++) {
+      if (measurementsNum != valueLists.get(i).length) {
+        throw new SQLParserException(
+            String.format(
+                "the measurementList's size %d is not consistent with the valueList's size %d",
+                measurementsNum, valueLists.get(i).length));
+      }
       InsertRowPlan insertRowPlan =
-          new InsertRowPlan(
-              device,
-              times[i],
-              measurementList,
-              Arrays.copyOfRange(valueList, i * measurementsNum, (i + 1) * measurementsNum));
+          new InsertRowPlan(device, times[i], measurementList, valueLists.get(i));
       insertRowPlan.setAligned(isAligned);
       insertRowsPlan.addOneInsertRowPlan(insertRowPlan, i);
     }
